Make Babbage route matching more specific
- Match the path exactly for the known endpoints - the parameters are on the query string so the path is constant
- For the known endpoints that have variable paths, continue to match the prefix
- Extract matcher functions to a separate file
