+++ router/router.go
@@ -6,11 +6,9 @@ import (
 	"github.com/ONSdigital/dp-frontend-router/middleware/serverError"
 	dprequest "github.com/ONSdigital/dp-net/request"
 	"github.com/ONSdigital/log.go/log"
-	"github.com/gorilla/mux"
 	"github.com/gorilla/pat"
 	"github.com/justinas/alice"
 	"net/http"
-	"path/filepath"
 	"strings"
 )
 
@@ -107,47 +105,4 @@ func healthcheckHandler(hc func(w http.ResponseWriter, req *http.Request)) func(
 			h.ServeHTTP(w, req)
 		})
 	}
-}
-
-var knownBabbageEndpoints = []string{
-	"/chartconfig",
-	"/chartimage",
-	"/generator",
-	"/visualisations/",
-	"/timeseriestool",
-	"/search",
-	"/file",
-	"/resource",
-	"/ons/",
-	"/file",
-	"/calendar",
-	"/chart",
-	"/embed",
-	"/export",
-	"/hash",
-}
-
-// IsKnownBabbageEndpoint returns true if the given path matches a known babbage endpoint
-func IsKnownBabbageEndpoint(path string) bool {
-	for _, endpoint := range knownBabbageEndpoints {
-		if strings.HasPrefix(path, endpoint) {
-			return true
-		}
-	}
-	return false
-}
-
-// hasFileExtMatcher is a mux MatcherFunc implementation, allowing routes to be matched on having a file extension
-func isKnownBabbageEndpointMatcher(request *http.Request, match *mux.RouteMatch) bool {
-	return IsKnownBabbageEndpoint(request.URL.Path)
-}
-
-// HasFileExt returns true if the given path has a file extension
-func HasFileExt(path string) bool {
-	return len(filepath.Ext(path)) > 0
-}
-
-// hasFileExtMatcher is a mux MatcherFunc implementation, allowing routes to be matched on having a file extension
-func hasFileExtMatcher(request *http.Request, match *mux.RouteMatch) bool {
-	return HasFileExt(request.URL.Path)
 }
