+++ router/router_test.go
@@ -17,58 +17,6 @@ func NewHandlerMock() *routertest.HandlerMock {
 	}
 }
 
-func TestHasFileExt(t *testing.T) {
-
-	Convey("Given a URL that has a file extension", t, func() {
-		Convey("When the HasFileExt function is called", func() {
-			Convey("Then true is returned", func() {
-				So(router.HasFileExt("/some/path.json"), ShouldBeTrue)
-				So(router.HasFileExt("/main.css"), ShouldBeTrue)
-				So(router.HasFileExt("/the/main/site.js"), ShouldBeTrue)
-			})
-		})
-	})
-
-	Convey("Given a URL that has no file extension", t, func() {
-		Convey("When the HasFileExt function is called", func() {
-			Convey("Then false is returned", func() {
-				So(router.HasFileExt("/some/path"), ShouldBeFalse)
-				So(router.HasFileExt("/"), ShouldBeFalse)
-				So(router.HasFileExt("/test/module.test/data"), ShouldBeFalse)
-			})
-		})
-	})
-}
-
-func TestIsKnownBabbageEndpoint(t *testing.T) {
-
-	Convey("Given a URL that is for a known babbage endpoint", t, func() {
-		Convey("When the IsKnownBabbageEndpoint function is called", func() {
-			Convey("Then true is returned", func() {
-				So(router.IsKnownBabbageEndpoint("/chartimage"), ShouldBeTrue)
-				So(router.IsKnownBabbageEndpoint("/visualisations/the/name"), ShouldBeTrue)
-				So(router.IsKnownBabbageEndpoint("/generator"), ShouldBeTrue)
-				So(router.IsKnownBabbageEndpoint("/chartconfig"), ShouldBeTrue)
-				So(router.IsKnownBabbageEndpoint("/search"), ShouldBeTrue)
-				So(router.IsKnownBabbageEndpoint("/resource"), ShouldBeTrue)
-				So(router.IsKnownBabbageEndpoint("/file"), ShouldBeTrue)
-				So(router.IsKnownBabbageEndpoint("/ons/some/more/url"), ShouldBeTrue)
-				So(router.IsKnownBabbageEndpoint("/timeseriestool"), ShouldBeTrue)
-				So(router.IsKnownBabbageEndpoint("/file"), ShouldBeTrue)
-			})
-		})
-	})
-
-	Convey("Given a URL that is not for a known babbage endpoint", t, func() {
-		Convey("When the IsKnownBabbageEndpoint function is called", func() {
-			Convey("Then false is returned", func() {
-				So(router.IsKnownBabbageEndpoint("/some/path"), ShouldBeFalse)
-				So(router.IsKnownBabbageEndpoint("/"), ShouldBeFalse)
-			})
-		})
-	})
-}
-
 func TestRouter(t *testing.T) {
 
 	Convey("Given a configured router", t, func() {
@@ -409,9 +357,47 @@ func TestRouter(t *testing.T) {
 			})
 		})
 
-		Convey("When a request for a known Babbage endpoint is made", func() {
+		Convey("When a request for a visualisation endpoint is made", func() {
 
 			url := "/visualisations/dvc1119"
+			req := httptest.NewRequest("GET", url, nil)
+			res := httptest.NewRecorder()
+
+			router := router.New(config)
+			router.ServeHTTP(res, req)
+
+			Convey("Then a request is not sent to Zebedee to check the page type", func() {
+				So(len(zebedeeClient.GetWithHeadersCalls()), ShouldEqual, 0)
+			})
+
+			Convey("Then the request is sent to the babbage", func() {
+				So(len(babbageHandler.ServeHTTPCalls()), ShouldEqual, 1)
+				So(babbageHandler.ServeHTTPCalls()[0].In2.URL.Path, ShouldResemble, url)
+			})
+		})
+
+		Convey("When a request for a legacy ons redirect endpoint is made", func() {
+
+			url := "/ons/some/old/page"
+			req := httptest.NewRequest("GET", url, nil)
+			res := httptest.NewRecorder()
+
+			router := router.New(config)
+			router.ServeHTTP(res, req)
+
+			Convey("Then a request is not sent to Zebedee to check the page type", func() {
+				So(len(zebedeeClient.GetWithHeadersCalls()), ShouldEqual, 0)
+			})
+
+			Convey("Then the request is sent to the babbage", func() {
+				So(len(babbageHandler.ServeHTTPCalls()), ShouldEqual, 1)
+				So(babbageHandler.ServeHTTPCalls()[0].In2.URL.Path, ShouldResemble, url)
+			})
+		})
+
+		Convey("When a request for a known babbage endpoint is made", func() {
+
+			url := "/file"
 			req := httptest.NewRequest("GET", url, nil)
 			res := httptest.NewRecorder()
